backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
duckdb
psycopg2-binary
mysql-connector-python
plotly
pandas
tqdm
google-generativeai